requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
codegen==1.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.3
django-rest-knox==4.1.0
djangorestframework==3.12.4
future==0.18.2
idna==2.10
npm==0.1.1
oauthlib==3.1.0
optional-django==0.1.0
patterns==0.3
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
serializers==0.2.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
urllib3==1.26.4
djongo==1.3.6
django-cors-headers==3.7.0